Add create_service.rb
